Create UpdateHolidaysCalendar.ps1
Script to update Holidays for all employees. It works with EXO and Exchange Server. It allows to work with 2 different locations (ARG and USA in the example) and to set FreeBusy accordingly to each location. This lets employees know about each location holidays in advance but it doesn't set OOF status for the location they don't reside in. Instead the holiday is added as an AllDay event for information.